+++ packages/next/src/views/Root/index.tsx
@@ -14,7 +14,7 @@ import { PageConfigProvider } from '@payloadcms/ui'
 import { RenderServerComponent } from '@payloadcms/ui/elements/RenderServerComponent'
 import { getClientConfig } from '@payloadcms/ui/utilities/getClientConfig'
 import { notFound, redirect } from 'next/navigation.js'
-import { formatAdminURL } from 'payload/shared'
+import { applyLocaleFiltering, formatAdminURL } from 'payload/shared'
 import * as qs from 'qs-esm'
 import React from 'react'
 
@@ -239,13 +239,13 @@ export const RootPage = async ({
     redirect(adminRoute)
   }
 
-  const clientConfig = await getClientConfig({
+  const clientConfig = getClientConfig({
     config,
     i18n: req.i18n,
     importMap,
-    req,
     user: viewType === 'createFirstUser' ? true : req.user,
   })
+  await applyLocaleFiltering({ clientConfig, config, req })
 
   const visibleEntities = getVisibleEntities({ req })
