+++ packages/ui/src/utilities/getClientConfig.ts
@@ -13,7 +13,7 @@ if (!cachedClientConfigs) {
 }
 
 export const getClientConfig = cache(
-  async ({ config, i18n, importMap, req, user }: CreateClientConfigArgs): Promise<ClientConfig> => {
+  ({ config, i18n, importMap, user }: CreateClientConfigArgs): ClientConfig => {
     const currentLanguage = i18n.language
 
     if (cachedClientConfigs[currentLanguage] && !global._payload_doNotCacheClientConfig) {
@@ -23,32 +23,15 @@ export const getClientConfig = cache(
         }) as unknown as ClientConfig
       }
 
-      const clientConfig = cachedClientConfigs[currentLanguage]
-
-      const configWithFilteredLocales = await applyLocaleFiltering({ clientConfig, config, req })
-      if (configWithFilteredLocales) {
-        return configWithFilteredLocales
-      }
-
-      return clientConfig
+      return cachedClientConfigs[currentLanguage]
     }
 
     const cachedClientConfig = createClientConfig({
       config,
       i18n,
       importMap,
-      req,
       user,
     })
-
-    const configWithFilteredLocales = await applyLocaleFiltering({
-      clientConfig: cachedClientConfig,
-      config,
-      req,
-    })
-    if (configWithFilteredLocales) {
-      Object.assign(cachedClientConfig, configWithFilteredLocales)
-    }
 
     cachedClientConfigs[currentLanguage] = cachedClientConfig
     global._payload_clientConfigs = cachedClientConfigs
@@ -63,37 +46,3 @@ export const getClientConfig = cache(
     return cachedClientConfig
   },
 )
-
-async function applyLocaleFiltering({
-  clientConfig,
-  config,
-  req,
-}: {
-  clientConfig: ClientConfig
-  config: CreateClientConfigArgs['config']
-  req: CreateClientConfigArgs['req']
-}): Promise<ClientConfig | null> {
-  if (
-    !clientConfig.localization ||
-    !config.localization ||
-    typeof config.localization.filterAvailableLocales !== 'function'
-  ) {
-    return null
-  }
-
-  const filteredLocales = (
-    await config.localization.filterAvailableLocales({
-      locales: config.localization.locales,
-      req,
-    })
-  ).map(({ toString, ...rest }) => rest)
-
-  return {
-    ...clientConfig,
-    localization: {
-      ...clientConfig.localization,
-      localeCodes: filteredLocales.map(({ code }) => code),
-      locales: filteredLocales,
-    },
-  }
-}
